# public/openapi/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api para a aplicação Maliin",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "Local"
        },
        {
            "url": "https://api.hml.malf88.xyz/api/",
            "description": "Homologação"
        }
    ],
    "paths": {
...
        "/account": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Retorna uma lista de contas",
                "description": "Retornará uma lista de cursos do usuário logado",
                "operationId": "d1d66a376984566eecf52a66018be6c3",
                "responses": {
                    "200": {
                        "description": "Uma lista de contas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Insere uma conta",
                "description": "Insere uma nova conta",
                "operationId": "c77feea4bf443d70328f72ebde2d22f7",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "bank": {
                                        "type": "string"
                                    },
                                    "agency": {
                                        "type": "string"
                                    },
                                    "account": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Nubank",
                                    "bank": "260",
                                    "account": "1234567-8",
                                    "agency": "0001"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Objeto inserido com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/account/{id}": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Retorna a conta com o {id}",
                "description": "Retorna a conta com o {id} informado",
                "operationId": "ac9cbe15f91cf8aa472535b129e34944",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do registro buscado",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Um objeto de curso"
                    },
                    "404": {
                        "description": "Conta não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Altera uma conta",
                "description": "Altera uma conta existente",
                "operationId": "cdd06cef53bfadbdf9dfe19191451cf9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do registro buscado",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "bank": {
                                        "type": "string"
                                    },
                                    "agency": {
                                        "type": "string"
                                    },
                                    "account": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Nubank",
                                    "bank": "260",
                                    "acount": "1234567-8",
                                    "agency": "0001"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Objeto alterado com sucesso"
                    },
                    "404": {
                        "description": "Objeto não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Exclui a conta com o {id}",
                "description": "Exclui a conta com o {id} informado",
                "operationId": "ede4536aa575fcdab4dca746e073f303",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do registro buscado",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excluído com sucesso"
                    },
                    "404": {
                        "description": "Conta não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/account/{account_id}/user/{user_id}": {
            "put": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Adiciona um usuário a conta",
                "description": "Adiciona um usuário a conta",
                "operationId": "cf65fc96fec259364c5be415cb8b10e3",
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "Id da conta",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Id do usuário a ser vinculado à conta",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excluído com sucesso"
                    },
                    "404": {
                        "description": "Conta não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Exclui o usuário da conta",
                "description": "Exclui o usuário da conta",
                "operationId": "d4c70047ab9e3e7089e37e4aa80a5bea",
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "Id da conta",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Id do usuário a ser removido da conta",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excluído com sucesso"
                    },
                    "404": {
                        "description": "Conta não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/account/{id}/user": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Retorna uma lista de usuários com acesso a conta.",
                "description": "Retorna lista de usuários com acesso a conta",
                "operationId": "65776fd8091dc9ff6ccc5d3b71d22533",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da conta",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Um objeto de curso"
                    },
                    "404": {
                        "description": "Conta não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "token"
            }
        }
    }
}